split - reuse Microphone component
